Add pages/signup dir
