+++ components/pages/login/LoginFailedToastContents.tsx
@@ -1,0 +1,11 @@
+const LoginFailedToastContents = () => {
+  return (
+    <>
+      Login failed.
+      <br />
+      Please check details and try again.
+    </>
+  );
+};
+
+export default LoginFailedToastContents;
