+++ components/pages/signup/SignUpFailedToastContents.tsx
@@ -1,0 +1,11 @@
+const SignUpFailedToastContents = () => {
+  return (
+    <>
+      Sign up failed.
+      <br />
+      Please check details and try again.
+    </>
+  );
+};
+
+export default SignUpFailedToastContents;
